backend/pyproject.toml
[project]
name = "face-analysis-backend"
version = "0.1.0"
description = "InsightFace-powered face analysis API"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "insightface>=0.7.3",
    "onnxruntime-gpu==1.19.2",
    "numpy<2",
    "opencv-python-headless>=4.9",
    "pillow>=10",
    "pydantic>=2",
    # onnxruntime-gpu 1.19 requires cuDNN 9.x; pin to a version with linux-x86_64 wheels
    "nvidia-cudnn-cu12>=9.1,<9.20",
    "nvidia-cublas-cu12>=12.4",
]

[tool.uv]
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
